+++ core/frontend/src/test/render/webgl/SurfaceTransparency.test.ts
@@ -17,7 +17,8 @@ import { ScreenViewport } from "../../../Viewport";
 import { Target } from "../../../render/webgl/Target";
 import { Primitive } from "../../../render/webgl/Primitive";
 import { RenderPass } from "../../../render/webgl/RenderFlags";
-import { MeshGraphic, SurfaceGeometry } from "../../../render/webgl/Mesh";
+import { MeshGraphic } from "../../../render/webgl/Mesh";
+import { SurfaceGeometry } from "../../../render/webgl/SurfaceGeometry";
 import { MeshArgs } from "../../../render/primitives/mesh/MeshPrimitives";
 import { MeshParams } from "../../../render/primitives/VertexTable";
 import { createBlankConnection } from "../../createBlankConnection";
