+++ core/frontend/src/webgl.ts
@@ -20,6 +20,7 @@ export * from "./render/webgl/ClipVolume";
 export * from "./render/webgl/ColorInfo";
 export * from "./render/webgl/Diagnostics";
 export * from "./render/webgl/DrawCommand";
+export * from "./render/webgl/EdgeGeometry";
 export * from "./render/webgl/EdgeSettings";
 export * from "./render/webgl/FeatureOverrides";
 export * from "./render/webgl/FloatRGBA";
@@ -43,6 +44,7 @@ export * from "./render/webgl/ScreenSpaceEffect";
 export * from "./render/webgl/ShaderBuilder";
 export * from "./render/webgl/ShaderProgram";
 export * from "./render/webgl/ShadowUniforms";
+export * from "./render/webgl/SurfaceGeometry";
 export * from "./render/webgl/Sync";
 export * from "./render/webgl/System";
 export * from "./render/webgl/PerformanceMetrics";
